feat: created writePackageJSON file
