+++ helpers/CreateNextTailwind.ts
@@ -6,6 +6,7 @@ import { execSync } from "child_process";
 import { ProjectOptions } from "../types";
 import { setupPrisma } from "./setupPrisma";
 import { PKG_ROOT } from "../consts";
+import { writePackageJSON } from "./writePackageJSON";
 
 async function createNextTailwindBoilerplate(
 	projectPath: string,
@@ -27,8 +28,8 @@ async function createNextTailwindBoilerplate(
 			message: "Template files copied with Hard-Coded Authentication.",
 		},
 		None: {
-			templatePath: null, // No additional template required
-			prismasrcpath: null,
+			templatePath: "templates/next-tailwind",
+			prismasrcpath: "base.prisma",
 			message: "Template files copied with No Authentication.",
 		},
 	};
@@ -56,51 +57,36 @@ async function createNextTailwindBoilerplate(
 		// this is for package.json setup
 		if (auth.templatePath) {
 			if (authentication === "NextAuth") {
-				const packageJsonPath = path.join(projectPath, "package.json");
-				const packageJson = await fs.readJson(packageJsonPath);
-
-				packageJson.dependencies = {
-					...packageJson.dependencies,
-					"next-auth": "4.24.10",
-					bcrypt: "5.1.0",
-				};
-				packageJson.devDependencies = {
-					...packageJson.devDependencies,
-					"@types/bcrypt": "5.0.0",
-				};
-
-				await fs.writeJson(packageJsonPath, packageJson, { spaces: 2 });
+				writePackageJSON(
+					projectPath,
+					{ "next-auth": "4.24.10", bcrypt: "5.1.0" },
+					{ "@types/bcrypt": "5.0.0" }
+				);
 			}
 
 			if (authentication === "Hard-coded") {
-				const packageJsonPath = path.join(projectPath, "package.json");
-				const packageJson = await fs.readJson(packageJsonPath);
-
-				packageJson.dependencies = {
-					...packageJson.dependencies,
-					bcryptjs: "^2.4.3",
-					"@nextui-org/react": "^2.2.10",
-					"@radix-ui/react-avatar": "^1.0.4",
-					"@radix-ui/react-dialog": "^1.0.5",
-					"@radix-ui/react-dropdown-menu": "^2.0.6",
-					"@radix-ui/react-label": "^2.0.2",
-					"@radix-ui/react-scroll-area": "^1.0.5",
-					"@radix-ui/react-slot": "^1.0.2",
-					axios: "^1.6.8",
-					"react-hot-toast": "^2.4.1",
-					"tailwindcss-animate": "^1.0.7",
-					"class-variance-authority": "^0.7.0",
-					"framer-motion": "^11.0.22",
-					jsonwebtoken: "^9.0.2",
-					"lucide-react": "^0.365.0",
-					nodemailer: "^6.9.12",
-				};
-				packageJson.devDependencies = {
-					...packageJson.devDependencies,
-					"@types/jsonwebtoken": "^9.0.6",
-				};
-
-				await fs.writeJson(packageJsonPath, packageJson, { spaces: 2 });
+				writePackageJSON(
+					projectPath,
+					{
+						bcryptjs: "^2.4.3",
+						"@nextui-org/react": "^2.2.10",
+						"@radix-ui/react-avatar": "^1.0.4",
+						"@radix-ui/react-dialog": "^1.0.5",
+						"@radix-ui/react-dropdown-menu": "^2.0.6",
+						"@radix-ui/react-label": "^2.0.2",
+						"@radix-ui/react-scroll-area": "^1.0.5",
+						"@radix-ui/react-slot": "^1.0.2",
+						axios: "^1.6.8",
+						"react-hot-toast": "^2.4.1",
+						"tailwindcss-animate": "^1.0.7",
+						"class-variance-authority": "^0.7.0",
+						"framer-motion": "^11.0.22",
+						jsonwebtoken: "^9.0.2",
+						"lucide-react": "^0.365.0",
+						nodemailer: "^6.9.12",
+					},
+					{ "@types/jsonwebtoken": "^9.0.6" }
+				);
 			}
 
 			await fs.copySync(
@@ -142,10 +128,6 @@ async function createNextTailwindBoilerplate(
 				)
 			);
 		}
-
-		console.log(
-			chalk.green("Next.js with Tailwind CSS boilerplate setup complete!")
-		);
 	} catch (error) {
 		spinner.fail(chalk.red("An error occurred during the setup process."));
 		console.error(error.message);
